fix(metrics): cleanup resource xds sync metrics
