+++ internal/kgateway/proxy_syncer/kube_gw_translator_syncer.go
@@ -3,7 +3,6 @@ package proxy_syncer
 import (
 	"context"
 
-	tmetrics "github.com/kgateway-dev/kgateway/v2/internal/kgateway/translator/metrics"
 	"github.com/kgateway-dev/kgateway/v2/pkg/logging"
 )
 
@@ -22,21 +21,10 @@ func (s *ProxyTranslator) syncXds(
 
 	logger.Log(ctx, logging.LevelTrace, "syncing xds snapshot", "proxy_key", proxyKey)
 
-	cd := getDetailsFromXDSClientResourceName(snapWrap.ResourceName())
-
 	// if the snapshot is not consistent, make it so
 	// TODO: me may need to copy this to not change krt cache.
 	// TODO: this is also may not be needed now that envoy has
 	// a default initial fetch timeout
 	// snap.MakeConsistent()
 	s.xdsCache.SetSnapshot(ctx, proxyKey, snap)
-
-	tmetrics.IncXDSSnapshotSync(cd.Gateway, cd.Namespace)
-
-	tmetrics.EndResourceSync(tmetrics.ResourceSyncDetails{
-		Namespace:    cd.Namespace,
-		Gateway:      cd.Gateway,
-		ResourceType: "XDSSnapshot",
-		ResourceName: cd.Gateway,
-	}, true, resourcesXDSSyncsTotal, resourcesXDSSyncDuration)
 }
