+++ internal/kgateway/proxy_syncer/status_syncer.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/kgateway-dev/kgateway/v2/api/v1alpha1"
 	"github.com/kgateway-dev/kgateway/v2/internal/kgateway/extensions2/common"
-	tmetrics "github.com/kgateway-dev/kgateway/v2/internal/kgateway/translator/metrics"
+	"github.com/kgateway-dev/kgateway/v2/internal/kgateway/krtcollections/metrics"
 	"github.com/kgateway-dev/kgateway/v2/internal/kgateway/utils"
 	"github.com/kgateway-dev/kgateway/v2/internal/kgateway/wellknown"
 	plug "github.com/kgateway-dev/kgateway/v2/pkg/pluginsdk"
@@ -194,12 +194,12 @@ func (s *StatusSyncer) syncRouteStatus(ctx context.Context, logger *slog.Logger,
 
 				defer func() {
 					for _, gatewayName := range gatewayNames {
-						tmetrics.EndResourceSync(tmetrics.ResourceSyncDetails{
+						metrics.EndResourceStatusSync(metrics.ResourceSyncDetails{
 							Namespace:    routeKey.Namespace,
 							Gateway:      gatewayName,
 							ResourceType: routeType,
 							ResourceName: routeKey.Name,
-						}, false, resourcesStatusSyncsCompletedTotal, resourcesStatusSyncDuration)
+						})
 
 						if finish, exists := finishMetrics[gatewayName]; exists {
 							finish.finishFunc(errors.Join(rErr, finish.statusError))
@@ -413,12 +413,12 @@ func (s *StatusSyncer) syncGatewayStatus(ctx context.Context, logger *slog.Logge
 		}
 
 		// Record metrics for this gateway
-		tmetrics.EndResourceSync(tmetrics.ResourceSyncDetails{
+		metrics.EndResourceStatusSync(metrics.ResourceSyncDetails{
 			Namespace:    gwnn.Namespace,
 			Gateway:      gwnn.Name,
 			ResourceType: wellknown.GatewayKind,
 			ResourceName: gwnn.Name,
-		}, false, resourcesStatusSyncsCompletedTotal, resourcesStatusSyncDuration)
+		})
 
 		finishMetrics(errors.Join(err, statusErr))
 	}
@@ -484,12 +484,12 @@ func (s *StatusSyncer) syncListenerSetStatus(ctx context.Context, logger *slog.L
 					logger.Debug("skipping k8s ls status update, status equal", "listenerset", lsnn.String())
 				}
 
-				tmetrics.EndResourceSync(tmetrics.ResourceSyncDetails{
+				metrics.EndResourceStatusSync(metrics.ResourceSyncDetails{
 					Namespace:    ls.Namespace,
 					Gateway:      string(ls.Spec.ParentRef.Name),
 					ResourceType: "XListenerSet",
 					ResourceName: ls.Name,
-				}, false, resourcesStatusSyncsCompletedTotal, resourcesStatusSyncDuration)
+				})
 			}
 		}
 		return nil
@@ -579,21 +579,12 @@ func (s *StatusSyncer) syncPolicyStatus(ctx context.Context, rm reports.ReportMa
 			continue
 		}
 
-		for _, ancestor := range status.Ancestors {
-			if ancestor.AncestorRef.Kind != nil && *ancestor.AncestorRef.Kind == "Gateway" {
-				namespace := nsName.Namespace
-				if ancestor.AncestorRef.Namespace != nil {
-					namespace = string(*ancestor.AncestorRef.Namespace)
-				}
-
-				tmetrics.EndResourceSync(tmetrics.ResourceSyncDetails{
-					Namespace:    namespace,
-					Gateway:      string(ancestor.AncestorRef.Name),
-					ResourceType: gk.Kind,
-					ResourceName: nsName.Name,
-				}, false, resourcesStatusSyncsCompletedTotal, resourcesStatusSyncDuration)
-			}
-		}
+		metrics.EndResourceStatusSync(metrics.ResourceSyncDetails{
+			Namespace:    nsName.Namespace,
+			Gateway:      "",
+			ResourceType: gk.Kind,
+			ResourceName: nsName.Name,
+		})
 
 		finishMetrics(statusErr)
 	}
