+++ internal/kgateway/translator/irtranslator/policy.go
@@ -8,7 +8,6 @@ import (
 	gwv1 "sigs.k8s.io/gateway-api/apis/v1"
 
 	"github.com/kgateway-dev/kgateway/v2/api/v1alpha1"
-	"github.com/kgateway-dev/kgateway/v2/internal/kgateway/translator/metrics"
 	"github.com/kgateway-dev/kgateway/v2/pkg/pluginsdk/ir"
 	"github.com/kgateway-dev/kgateway/v2/pkg/pluginsdk/reporter"
 )
@@ -25,13 +24,6 @@ func reportPolicyAcceptanceStatus(
 			// Not a policy associated with a CR, can't report status on it
 			continue
 		}
-
-		// Start the resource sync metrics for the policy.
-		metrics.StartResourceSync(policy.PolicyRef.Name, metrics.ResourceMetricLabels{
-			Gateway:   string(ancestorRef.Name),
-			Namespace: policy.PolicyRef.Namespace,
-			Resource:  policy.PolicyRef.Kind,
-		})
 
 		key := reporter.PolicyKey{
 			Group:     policy.PolicyRef.Group,
