+++ test/kubernetes/e2e/features/metrics/suite.go
@@ -84,26 +84,20 @@ func (s *testingSuite) TestMetrics() {
 				gomega.MatchRegexp(`kgateway_resources_managed\{namespace=\"default\",parent=\"gw2\",resource=\"Gateway\"} 1`),
 				gomega.MatchRegexp(`kgateway_resources_managed\{namespace=\"default\",parent=\"gw2\",resource=\"HTTPRoute\"} 1`),
 				gomega.MatchRegexp(`kgateway_resources_managed\{namespace=\"default\",parent=\"ls1\",resource=\"HTTPRoute\"} 1`),
-				gomega.MatchRegexp(`kgateway_resources_managed\{namespace=\"default\",parent=\"gw2\",resource=\"HTTPListenerPolicy\"} 1`),
-				gomega.MatchRegexp(`kgateway_resources_syncs_started_total\{gateway=\"gw2\",namespace=\"default\",resource=\"Gateway\"} [2-6]`),
-				gomega.MatchRegexp(`kgateway_resources_syncs_started_total\{gateway=\"gw2\",namespace=\"default\",resource=\"HTTPRoute\"} [2-3]`),
-				gomega.MatchRegexp(`kgateway_resources_syncs_started_total\{gateway=\"gw2\",namespace=\"default\",resource=\"HTTPListenerPolicy\"} [1-2]`),
+				gomega.MatchRegexp(`kgateway_resources_managed\{namespace=\"default\",parent=\"\",resource=\"HTTPListenerPolicy\"} 1`),
+				gomega.MatchRegexp(`kgateway_resources_status_syncs_started_total\{gateway=\"gw2\",namespace=\"default\",resource=\"Gateway\"} [2-6]`),
+				gomega.MatchRegexp(`kgateway_resources_status_syncs_started_total\{gateway=\"gw2\",namespace=\"default\",resource=\"HTTPRoute\"} [1-3]`),
+				gomega.MatchRegexp(`kgateway_resources_status_syncs_started_total\{gateway=\"\",namespace=\"default\",resource=\"HTTPListenerPolicy\"} [1-2]`),
 				gomega.MatchRegexp(`kgateway_resources_status_syncs_completed_total\{gateway=\"gw2\",namespace=\"default\",resource=\"Gateway\"} [2-6]`),
-				gomega.MatchRegexp(`kgateway_resources_status_syncs_completed_total\{gateway=\"gw2\",namespace=\"default\",resource=\"HTTPRoute\"} [2-3]`),
-				gomega.MatchRegexp(`kgateway_resources_status_syncs_completed_total\{gateway=\"gw2\",namespace=\"default\",resource=\"HTTPListenerPolicy\"} [1-2]`),
+				gomega.MatchRegexp(`kgateway_resources_status_syncs_completed_total\{gateway=\"gw2\",namespace=\"default\",resource=\"HTTPRoute\"} [1-3]`),
+				gomega.MatchRegexp(`kgateway_resources_status_syncs_completed_total\{gateway=\"\",namespace=\"default\",resource=\"HTTPListenerPolicy\"} [1-2]`),
 				gomega.MatchRegexp(`kgateway_resources_status_sync_duration_seconds_count\{gateway=\"gw2\",namespace=\"default\",resource=\"Gateway\"} [2-6]`),
-				gomega.MatchRegexp(`kgateway_resources_status_sync_duration_seconds_count\{gateway=\"gw2\",namespace=\"default\",resource=\"HTTPRoute\"} [2-3]`),
-				gomega.MatchRegexp(`kgateway_resources_status_sync_duration_seconds_count\{gateway=\"gw2\",namespace=\"default\",resource=\"HTTPListenerPolicy\"} [1-2]`),
-				gomega.MatchRegexp(`kgateway_resources_xds_snapshot_syncs_total\{gateway=\"gw1\",namespace=\"default\",resource=\"Gateway\"} 1`),
-				gomega.MatchRegexp(`kgateway_resources_xds_snapshot_syncs_total\{gateway=\"gw2\",namespace=\"default\",resource=\"Gateway\"} 1`),
-				gomega.MatchRegexp(`kgateway_resources_xds_snapshot_syncs_total\{gateway=\"gw2\",namespace=\"default\",resource=\"HTTPRoute\"} 1`),
-				gomega.MatchRegexp(`kgateway_resources_xds_snapshot_syncs_total\{gateway=\"gw2\",namespace=\"default\",resource=\"HTTPListenerPolicy\"} 1`),
-				gomega.MatchRegexp(`kgateway_resources_xds_snapshot_sync_duration_seconds_count\{gateway=\"gw1\",namespace=\"default\",resource=\"Gateway\"} 1`),
-				gomega.MatchRegexp(`kgateway_resources_xds_snapshot_sync_duration_seconds_count\{gateway=\"gw2\",namespace=\"default\",resource=\"Gateway\"} 1`),
-				gomega.MatchRegexp(`kgateway_resources_xds_snapshot_sync_duration_seconds_count\{gateway=\"gw2\",namespace=\"default\",resource=\"HTTPRoute\"} 1`),
-				gomega.MatchRegexp(`kgateway_resources_xds_snapshot_sync_duration_seconds_count\{gateway=\"gw2\",namespace=\"default\",resource=\"HTTPListenerPolicy\"} 1`),
-				gomega.MatchRegexp(`kgateway_xds_snapshot_syncs_total\{gateway=\"gw1\",namespace=\"default\"} \d+`),
-				gomega.MatchRegexp(`kgateway_xds_snapshot_syncs_total\{gateway=\"gw2\",namespace=\"default\"} \d+`),
+				gomega.MatchRegexp(`kgateway_resources_status_sync_duration_seconds_count\{gateway=\"gw2\",namespace=\"default\",resource=\"HTTPRoute\"} [1-3]`),
+				gomega.MatchRegexp(`kgateway_resources_status_sync_duration_seconds_count\{gateway=\"\",namespace=\"default\",resource=\"HTTPListenerPolicy\"} [1-2]`),
+				gomega.MatchRegexp(`kgateway_xds_snapshot_syncs_total\{gateway=\"gw1\",namespace=\"default\"} [1-2]`),
+				gomega.MatchRegexp(`kgateway_xds_snapshot_syncs_total\{gateway=\"gw2\",namespace=\"default\"} [1-2]`),
+				gomega.MatchRegexp(`kgateway_xds_snapshot_sync_duration_seconds_count\{gateway=\"gw1\",namespace=\"default\"} [1-2]`),
+				gomega.MatchRegexp(`kgateway_xds_snapshot_sync_duration_seconds_count\{gateway=\"gw2\",namespace=\"default\"} [1-2]`),
 				gomega.MatchRegexp(`kgateway_xds_snapshot_transform_duration_seconds_count\{gateway=\"gw1\",namespace=\"default\"} \d+`),
 				gomega.MatchRegexp(`kgateway_xds_snapshot_transforms_total\{gateway=\"gw1\",namespace=\"default\",result="success"} \d+`),
 				gomega.MatchRegexp(`kgateway_xds_snapshot_resources\{gateway=\"gw1\",namespace=\"default\",resource=\"Endpoint\"} \d+`),
